kicker/requirements.txt
certifi
cffi
charset-normalizer
idna
pycparser
PyNaCl
requests
urllib3
